Vulkan: Combine PaletteTextureConverter and TextureEncoder classes
