+++ Source/Core/VideoBackends/Vulkan/Vulkan.vcxproj
@@ -53,12 +53,11 @@
     <ClCompile Include="CommandBufferManager.cpp" />
     <ClCompile Include="FramebufferManager.cpp" />
     <ClCompile Include="main.cpp" />
-    <ClCompile Include="PaletteTextureConverter.cpp" />
+    <ClCompile Include="TextureConverter.cpp" />
     <ClCompile Include="PerfQuery.cpp" />
     <ClCompile Include="RasterFont.cpp" />
     <ClCompile Include="StagingBuffer.cpp" />
     <ClCompile Include="StagingTexture2D.cpp" />
-    <ClCompile Include="TextureEncoder.cpp" />
     <ClCompile Include="Util.cpp" />
     <ClCompile Include="VertexFormat.cpp" />
     <ClCompile Include="ObjectCache.cpp" />
@@ -78,11 +77,10 @@
     <ClInclude Include="CommandBufferManager.h" />
     <ClInclude Include="FramebufferManager.h" />
     <ClInclude Include="Constants.h" />
-    <ClInclude Include="PaletteTextureConverter.h" />
+    <ClInclude Include="TextureConverter.h" />
     <ClInclude Include="RasterFont.h" />
     <ClInclude Include="StagingBuffer.h" />
     <ClInclude Include="StagingTexture2D.h" />
-    <ClInclude Include="TextureEncoder.h" />
     <ClInclude Include="Util.h" />
     <ClInclude Include="VertexFormat.h" />
     <ClInclude Include="PerfQuery.h" />
